Do not attach to the caching session if it may fail
When asking for a cached file, if the remote returns 304 (Not modified)
all the requests but the first are going to fail because they are not
handling the errback. With this, we don't attach to sessions that are
checking if the remote has been modified and just return the cached
file.